added functionality

+++ scripts/1_trim-gap-columns.sh
@@ -5,8 +5,11 @@ do
     do
     	echo "Working on" $alignment "in" $dir
         trimal -in $alignment -out $alignment.noallgaps -noallgaps -keepseqs
-        sed 's/ [0-9]* bp//g' $alignment.noallgaps > $alignment
-        rm $alignment.noallgaps
+        filename="$alignment.noallgaps"
+        if [ -e "$filename" ]; then
+            sed 's/ [0-9]* bp//g' $alignment.noallgaps > $alignment
+            rm $alignment.noallgaps
+        fi
     done
     cd ../../scripts/
 done
